requirements.txt
eclipse-zenoh
fastapi
uvicorn[standard]
pyyaml
aiortc
